docs(runtime): document execution topology
