+++ src/runtime/src/runtime/execution/module_environment.rs
@@ -1,4 +1,28 @@
-use super::*;
+use crate::context::{
+  RuntimeContextBinding,
+  RuntimeContextRegistry,
+};
+use crate::resolver::{
+  SourceExportDeclaration,
+  SourceImportAlias,
+  SourceImportDeclaration,
+  SourceImportKind,
+  SourceScope,
+};
+use crate::runtime::{
+  RuntimeInvalidOperationError,
+  RuntimeModuleImportConflict,
+};
+use crate::store::ModuleVersionRecord;
+use mech_core::{
+  hash_str,
+  MResult,
+  MechError,
+  MechSourceCode,
+  ValRef,
+};
+use mech_program::MechProgram;
+use std::collections::HashMap;
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub(super) enum RuntimeAddressTarget {
