+++ src/runtime/src/runtime/execution/query.rs
@@ -1,4 +1,14 @@
-use super::*;
+use crate::runtime::{
+  MechRuntime,
+  RuntimeInvalidOperationError,
+  RuntimeProgramBusy,
+};
+use crate::RuntimeValueSnapshot;
+use mech_core::{
+  MResult,
+  MechError,
+  Value,
+};
 
 impl MechRuntime {
   #[cfg(feature = "invariant_define")]
