+++ src/runtime/src/runtime/execution/source.rs
@@ -1,4 +1,34 @@
-use super::*;
+use super::{
+  identifier_from_str,
+  resolve_runtime_value,
+  single_code_program,
+  RuntimeProgramTarget,
+};
+use crate::event::RuntimeEventKind;
+use crate::resolver::SourceScope;
+use crate::runtime::{
+  MechRuntime,
+  RuntimeInvalidOperationError,
+  RuntimeProgramBusy,
+};
+use crate::{
+  ResourceBudgetExceededError,
+  RuntimeContext,
+  RuntimeValueSnapshot,
+};
+use mech_core::{
+  hash_str,
+  MResult,
+  MechError,
+  MechSourceCode,
+  ValRef,
+  Value,
+};
+use mech_program::MechProgram;
+#[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
+use web_time::Instant;
+#[cfg(not(all(target_arch = "wasm32", target_os = "unknown")))]
+use std::time::Instant;
 
 impl MechRuntime {
   pub(super) fn run_tree_on_program(
